requirements.txt
gradio~=3.23.0
docx2txt~=0.8
pdf2docx~=0.5.6